requirements.txt
azure-functions
requests
azure-core
azure-ai-formrecognizer